[WIP] feat(multi-client): Multi-client support for the application
With this change, helix becomes a multi-client application. `hx` will
automatically start the server process if necessary, and connect to it.
Buffers and language servers run as subprocesses of the server and are
used by all clients.

The client/server protocol is very simple. The server listens on a Unix
socket. The client opens the socket and sends a serialized ClientInfo to
the server, which contains the command line arguments, working directory
and some other information. The terminal file descriptor (and the stdin
file descriptor, if stdin is not a terminal) are then sent to the server
using SCM_RIGHTS. The server uses the tty file descriptor to render the
UI and receive user input exactly as it does today afer opening /dev/tty,
and uses the stdin file descriptor to create a scratch buffer.

The client listens for relevant signals such as SIGWINCH, SIGTSTP and
SIGCONT. When the client receives a signal, it sends a single byte signal
number to the server. The server responds to receiving a byte on the
socket similarly to how Helix handles signals today. One exception is
that in response to a SIGTSTP, the server sends SIGSTOP to the client's
process group instead of its own process group. This way, the server
keeps running while the client process (and the other processes in its
process group) is suspended.

When a client terminates (e.g. after :q), the server sends a single byte
exit code to the client, and the client exits using the exit code.

TODO:
- Unbreak the integration tests.
- Unbreak the Windows build. For the time being, the native Windows build will
  not support multi-client. Windows users who want multi-client can use WSL.
- Unbreak per-workspace .helix/config.toml support.
- Wait until Helix switches to Termina (#13307), and then reimplement this
  on top of Termina. For now this depends on a forked version of Crossterm which
  is available at: https://github.com/pcc/crossterm/tree/client-server-rebase
  From a brief look at the Termina code it seems easier to extend
  Termina to support this than it was to extend Crossterm.