feat(multi-client): Improve :q behavior
Previously :q would check for any modified document. This isn't useful
behavior for multi-client because there may be unrelated clients with
modified files that we don't want to deal with right now.

Upstream Kakoune doesn't check whether the buffer is modified on :q
[1]. This is technically correct because the server won't have lost the
contents of the buffer, but it isn't very useful behavior because external
commands would still use the old file contents. Therefore I made a change
in my fork that checks whether the current buffer is modified [2]. But I
still have problems with unsaved files if I make a change while navigating
using the LSP. Therefore, this change makes it so that :q checks whether
any of the views being closed have selections for an unsaved document,
which is roughly "files that were edited in this client".

[1] https://github.com/mawww/kakoune/blob/2d8c0b8bf0d7d18218d4c939003da39a018361e7/src/commands.cc#L782
[2] https://github.com/pcc/kakoune/commit/c31d4b6a29eaed6ebfd7f0e95668eeeefe1e55f1