[WIP] feat(multi-client): Make find_lsp_workspaces cwd-independent
Currently find_lsp_workspaces has some dependencies on the server's
current directory. Get rid of those and replace it with a different
implementation that just searches for a root marker in the ancestors.

This commit breaks the workspace-lsp-roots feature, which I don't use
and I'm not really sure how it's supposed to work anyway.